+++ src/main/java/com/coobird/staticlogistics/item/handler/FaceConfigModeHandler.java
@@ -23,6 +23,7 @@ public class FaceConfigModeHandler implements ModeHandler {
         var level = context.getLevel();
         var player = context.getPlayer();
         if (player == null) return InteractionResult.PASS;
+        if (!player.isSecondaryUseActive()) return InteractionResult.PASS;
         if (!level.isClientSide && player instanceof ServerPlayer serverPlayer && level instanceof ServerLevel serverLevel) {
             BlockPos pos = context.getClickedPos();
             Direction face = context.getClickedFace();